+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 
 - This version adds a new `AppScreenNavigationContext`.
 - This version adds a new `AppScreenNavigationStack`.
+- The `AppItem` type has less type-specific info.
 - The `AppItemAddContext` type is now `@MainActor`.